pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "work-buddy"
version = "0.1.0"
description = "Multi-agent system for automating SDLC workflows — Jira tasks, testing evidence, compliance validation, release prep, and production support"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Work Buddy Team"}
]

dependencies = [
    # Agent orchestration
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    # Browser automation
    "playwright>=1.48.0",
    # Jira & Confluence
    "jira>=3.8.0",
    "atlassian-python-api>=3.41.0",
    # HTTP client
    "httpx>=0.27.0",
    # CLI
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    # Configuration & validation
    "pyyaml>=6.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # RAG / Vector DB
    "chromadb>=0.5.0",
    # Mock services
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    # Utilities
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[project.scripts]
workbuddy = "work_buddy.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/work_buddy"]

[tool.hatch.build.targets.sdist]
include = ["src/work_buddy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
